trying to fix category normalisation

+++ clients/GVGAI-JavaClient/src/group/datagather/constants/Constants.java
@@ -8,9 +8,9 @@ public class Constants {
 	public static int CURRENT_GAME_ID;
 	public static int CURRENT_LEVEL_ID;
 
-	public static final int CATEGORY_NORMALISATION_FACTOR = 15;
-	public static final int NORMALISATION_FACTOR = 10000;
-	public static final int AVAILABLE_ACTIONS = 5;
+	public static final double CATEGORY_NORMALISATION_FACTOR = 15;
+	public static final double NORMALISATION_FACTOR = 10000;
+	public static final double AVAILABLE_ACTIONS = 5;
 	public static final int UP = 0, DOWN = 1, LEFT = 2, RIGHT = 3, USE = 4;
 	public static final int DV = 10;
 }
